refactor: Reorganize folder structure.
