+++ app/routes/_foh+/($championship)+/spiele/_matches.tsx
@@ -1,5 +1,5 @@
 import { type LoaderFunctionArgs, json } from '@remix-run/node';
-import { requireChampionship } from '#utils/foh/championships.server';
+import { requireChampionship } from '#utils/app/foh/championships.server';
 
 export async function loader({ params }: LoaderFunctionArgs) {
   const championship = await requireChampionship(params);
@@ -9,7 +9,7 @@ export async function loader({ params }: LoaderFunctionArgs) {
 export default function MatchesRoute() {
   return (
     <div>
-      <h2 className="text-xl font-medium">Spiele</h2>
+      <h2 className="font-medium text-xl">Spiele</h2>
     </div>
   );
 }
